// src/app/calendar/enums/parking-lot-button-style-enum.ts
export enum ParkingLotButtonStyleEnum {
  YOU_HAVE_PARKING_PLACE = 'YOU_HAVE_PARKING_PLACE',
  THERE_IS_FREE_PLACE = 'THERE_IS_FREE_PLACE',
  NOTHING_INTERESTING = 'NOTHING_INTERESTING',
  THERE_IS_FREE_PLACE_CLICKED = 'THERE_IS_FREE_PLACE_CLICKED',
  YOU_HAVE_PARKING_PLACE_CLICKED = 'YOU_HAVE_PARKING_PLACE_CLICKED',

}